+++ tests/integration.rs
@@ -4,9 +4,10 @@ use arrow::{
     array::{
         Array, ArrayRef, BinaryArray, BooleanArray, Date32Array, Date64Array, Decimal128Array,
         Decimal256Builder, FixedSizeBinaryArray, Float16Array, Float32Array, Int16Array,
-        Int32Array, Int64Array, Int8Array, StringArray, Time32MillisecondArray, Time32SecondArray,
-        Time64MicrosecondArray, Time64NanosecondArray, TimestampMicrosecondArray,
-        TimestampMillisecondArray, TimestampNanosecondArray, TimestampSecondArray, UInt8Array,
+        Int32Array, Int64Array, Int8Array, LargeStringArray, StringArray, Time32MillisecondArray,
+        Time32SecondArray, Time64MicrosecondArray, Time64NanosecondArray,
+        TimestampMicrosecondArray, TimestampMillisecondArray, TimestampNanosecondArray,
+        TimestampSecondArray, UInt8Array,
     },
     datatypes::{
         ArrowPrimitiveType, DataType, Decimal256Type, Field, Float16Type, Schema, SchemaRef,
@@ -1699,6 +1700,32 @@ fn insert_taking_ownership_of_connection() {
         OdbcWriter::from_connection(conn, &schema, table_name, row_capacity).unwrap()
     };
     inserter.write_all(reader).unwrap();
+
+    // Then
+    let actual = table_to_string(&conn, table_name, &["a"]);
+    let expected = "Hello\nNULL\nWorld";
+    assert_eq!(expected, actual);
+}
+
+#[test]
+fn insert_large_text() {
+    // Given a table and a record batch reader returning a batch with a text column.
+    let table_name = function_name!().rsplit_once(':').unwrap().1;
+    let conn = ENV
+        .connect_with_connection_string(MSSQL, Default::default())
+        .unwrap();
+    setup_empty_table(&conn, table_name, &["VARCHAR(4096)"]).unwrap();
+    let array = LargeStringArray::from(vec![Some("Hello"), None, Some("World")]);
+    let schema = Arc::new(Schema::new(vec![Field::new(
+        "a",
+        DataType::LargeUtf8,
+        true,
+    )]));
+    let batch = RecordBatch::try_new(schema.clone(), vec![Arc::new(array)]).unwrap();
+    let mut reader = StubBatchReader::new(schema, vec![batch]);
+
+    // When
+    insert_into_table(&conn, &mut reader, table_name, 5).unwrap();
 
     // Then
     let actual = table_to_string(&conn, table_name, &["a"]);